+++ spec/js_routes/default_serializer_spec.rb
@@ -17,4 +17,15 @@ describe JsRoutes, "#serialize" do
       "a=1&b%5B%5D=2&b%5B%5D=3&c%5Bd%5D=4&c%5Be%5D=5&f="
     )
   end
+
+  it "works with JS suckiness" do
+    expect(evaljs(
+      [
+        "const query = Object.create(null);",
+        "query.a = 1;",
+        "query.b = 2;",
+        "Routes.serialize(query);",
+      ].join("\n")
+    )).to eq("a=1&b=2")
+  end
 end